Converts sentinel lineage nodeunit tests
