--- sentinel/test/unit/lineage.js
@@ -1,753 +1,0 @@
-const db = require('../../db'),
-      utils = require('../../lib/utils'),
-      sinon = require('sinon').sandbox.create(),
-      lineage = require('../../lib/lineage');
-
-exports.tearDown = callback => {
-  sinon.restore();
-  callback();
-};
-
-let report_parentContact,
-  report_grandparentContact,
-  report_grandparent,
-  report_parent,
-  report_contact,
-  report_patient,
-  report,
-  place_parentContact,
-  place_grandparentContact,
-  place_contact,
-  place_parent,
-  place_grandparent,
-  place;
-
-exports.setUp = callback => {
-  report_parentContact = {
-    _id: 'report_parentContact',
-    name: 'report_parentContact_name',
-    type: 'person',
-    phone: '+123'
-  };
-  report_grandparentContact = {
-    _id: 'report_grandparentContact',
-    name: 'report_grandparentContact_name',
-    type: 'person',
-    phone: '+456'
-  };
-  report_grandparent = {
-    _id: 'report_grandparent',
-    contact: { _id: report_grandparentContact._id },
-    name: 'report_grandparent_name'
-  };
-  report_parent = {
-    _id: 'report_parent',
-    name: 'report_parent_name',
-    contact: { _id: report_parentContact._id },
-    parent: {
-      _id: report_grandparent._id
-    }
-  };
-  report_contact = {
-    _id: 'report_contact',
-    type: 'person',
-    name: 'report_contact_name',
-    parent: {
-      _id: report_parent._id,
-      parent: {
-        _id: report_grandparent._id
-      }
-    }
-  };
-  report_patient = {
-    _id: 'report-patient',
-    type: 'person',
-    name: 'patient_name',
-    parent: {
-      _id: report_contact._id,
-      parent: {
-        _id: report_parent._id,
-        parent: {
-          _id: report_grandparent._id
-        }
-      }
-    }
-  };
-  report = {
-    _id: 'report',
-    type: 'data_record',
-    form: 'A',
-    contact: {
-      _id: report_contact._id,
-      parent: {
-        _id: report_parent._id,
-        parent: {
-          _id: report_grandparent._id
-        }
-      }
-    },
-    fields: {
-      patient_id: 'abc123'
-    }
-  };
-
-  place_parentContact = {
-    _id: 'place_parentContact',
-    name: 'place_parentContact_name',
-    type: 'person',
-    phone: '+123'
-  };
-  place_grandparentContact = {
-    _id: 'place_grandparentContact',
-    name: 'place_grandparentContact_name',
-    phone: '+456'
-  };
-  place_contact = {
-    _id: 'place_contact',
-    type: 'place_contact_name',
-    phone: '+789'
-  };
-  place_grandparent = {
-    _id: 'place_grandparent',
-    name: 'place_grandparent_name',
-    contact: { _id: place_grandparentContact._id }
-  };
-  place_parent = {
-    _id: 'place_parent',
-    name: 'place_parent_name',
-    contact: { _id: place_parentContact._id },
-    parent: {
-      _id: place_grandparent._id
-    }
-  };
-  place = {
-    _id: 'place',
-    name: 'place_name',
-    contact: { _id: place_contact._id },
-    parent: {
-      _id: place_parent._id,
-      parent: {
-        _id: place_grandparent._id
-      }
-    }
-  };
-  callback();
-};
-
-exports['fetchHydratedDoc returns errors from view'] = test => {
-  const expected = 'boom';
-  const id = 'abc';
-  const view = sinon.stub(db.medic, 'view').callsArgWith(3, expected);
-  lineage.fetchHydratedDoc(id).catch(err => {
-    test.equals(err, expected);
-    test.equals(view.callCount, 1);
-    test.equals(view.args[0][0], 'medic-client');
-    test.equals(view.args[0][1], 'docs_by_id_lineage');
-    test.deepEqual(view.args[0][2].startkey, [ id ]);
-    test.deepEqual(view.args[0][2].endkey, [ id, {} ]);
-    test.deepEqual(view.args[0][2].include_docs, true);
-    test.done();
-  });
-};
-
-exports['fetchHydratedDoc returns errors from fetch'] = test => {
-  const expected = 'boom';
-  const docId = 'abc';
-  const contactId = 'def';
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, {
-    rows: [ { doc: { contact: { _id: contactId } } } ]
-  });
-  const fetch = sinon.stub(db.medic, 'fetch').callsArgWith(1, expected);
-  lineage.fetchHydratedDoc(docId).catch(err => {
-    test.equals(err, expected);
-    test.equals(fetch.callCount, 1);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc handles no lineage and no contact'] = test => {
-  const docId = 'abc';
-  const expected = { _id: docId };
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [ { doc: expected } ] });
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [] });
-  lineage.fetchHydratedDoc(docId).then(actual => {
-    test.deepEqual(actual, expected);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc handles non-lineage types with empty lineages'] = test => {
-  const expected = {_id: 'blah-info'};
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [] });
-  sinon.stub(db.medic, 'get').callsArgWith(1, null, expected);
-  lineage.fetchHydratedDoc(expected._id).then(actual => {
-    test.deepEqual(actual, expected);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc handles doc with unknown parent by leaving just the stub'] = test => {
-  const docId = 'abc';
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: { _id: 'abc', contact: { _id: 'def' }, parent: { _id: 'ghi', parent: { _id: 'jkl' } } } },
-    { doc: null },
-    { doc: { _id: 'jkl', name: 'district' } }
-  ] });
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [
-    { doc: { _id: 'cba' } },
-    { doc: { _id: 'def' } }
-  ] });
-  lineage.fetchHydratedDoc(docId)
-  .then(actual => {
-    test.deepEqual(actual, {
-      _id: 'abc',
-      contact: { _id: 'def' },
-      parent: { _id: 'ghi', parent: { _id: 'jkl', name: 'district' } }
-    });
-    test.done();
-  })
-  .catch(err => {
-    test.fail(err);
-    test.done();
-  });
-};
-
-exports['fetchHydratedDoc handles doc with unknown contact by leaving just the stub'] = test => {
-  const docId = 'abc';
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: { _id: 'abc', type: 'data_record', contact: { _id: 'def', parent: { _id: 'ghi', parent: { _id: 'jkl' } } } } },
-    { doc: null },
-    { doc: { _id: 'ghi', name: 'clinic' } },
-    { doc: { _id: 'jkl', name: 'district' } }
-  ] });
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [
-    { doc: { _id: 'cba' } },
-    { doc: { _id: 'def', parent: { _id: 'ghi', parent: { _id: 'jkl' } } } }
-  ] });
-  lineage.fetchHydratedDoc(docId)
-    .then(actual => {
-      test.deepEqual(actual, {
-        _id: 'abc',
-        type: 'data_record',
-        contact: {
-          _id: 'def',
-          parent: {
-            _id: 'ghi',
-            name: 'clinic',
-            parent: {
-              _id: 'jkl',
-              name: 'district'
-            }
-          }
-        }
-      });
-      test.done();
-    })
-    .catch(err => {
-      test.fail(err);
-      test.done();
-    });
-};
-
-exports['fetchHydratedDoc handles missing contacts'] = test => {
-  const docId = 'abc';
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [ { doc: { _id: 'abc', contact: { _id: 'xyz' } } } ] });
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [ { doc: null }] });
-  lineage.fetchHydratedDoc(docId).then(actual => {
-    test.deepEqual(actual, { _id: 'abc', contact: { _id: 'xyz' } });
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc attaches the lineage'] = test => {
-  const docId = 'abc';
-  const parent = {
-    _id: 'def',
-    name: 'jack',
-    parent: {
-      _id: 'ghi'
-    }
-  };
-  const grandparent = {
-    _id: 'ghi',
-    name: 'jim'
-  };
-  const given = {
-    _id: docId,
-    name: 'joe',
-    parent: {
-      _id: parent._id,
-      parent: {
-        _id: grandparent._id
-      }
-    }
-  };
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: given },
-    { doc: parent },
-    { doc: grandparent }
-  ] });
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [] }); // without subcontacts
-  lineage.fetchHydratedDoc(docId).then(actual => {
-    test.equals(actual.name, 'joe');
-    test.equals(actual.parent.name, 'jack');
-    test.equals(actual.parent.parent.name, 'jim');
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc attaches the full lineage for reports'] = test => {
-  sinon.stub(utils, 'getPatientContactUuid').callsArgWith(2, null, report_patient._id);
-  const viewStub = sinon.stub(db.medic, 'view');
-  viewStub.onCall(0).callsArgWith(3, null, { rows: [
-    { doc: report },
-    { doc: report_contact },
-    { doc: report_parent },
-    { doc: report_grandparent }
-  ] });
-  viewStub.onCall(1).callsArgWith(3, null, { rows: [
-    { doc: report_patient },
-    { doc: report_contact },
-    { doc: report_parent },
-    { doc: report_grandparent }
-  ] });
-  viewStub.onCall(2).callsArgWith(3, 'Too many calls!');
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.callsFake((options, callback) => {
-    const contactDocs = options.keys.map(id => {
-      return [ report_contact, report_parentContact, report_grandparentContact, report_patient ]
-        .find(contact => contact._id === id);
-    });
-    callback(null, { rows: contactDocs.map(doc => { return {doc: doc}; }) });
-  });
-
-  lineage.fetchHydratedDoc(report._id).then(actual => {
-    test.equals(actual.form, 'A');
-    // parents
-    test.equals(actual.contact.name, report_contact.name);
-    test.equals(actual.contact.parent.name, report_parent.name);
-    test.equals(actual.contact.parent.parent.name, report_grandparent.name);
-    test.ok(!actual.parent);
-    // contacts of parents
-    test.equals(actual.contact.parent.contact.phone, '+123');
-    test.equals(actual.contact.parent.parent.contact.phone, '+456');
-    // patient
-    test.equals(actual.patient.name, report_patient.name);
-    // patient parents
-    test.equals(actual.patient.parent.name, report_contact.name);
-
-    test.equals(fetch.callCount, 1);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc attaches the contacts'] = test => {
-  const view = sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: place },
-    { doc: place_parent },
-    { doc: place_grandparent }
-  ] });
-  const fetch = sinon.stub(db.medic, 'fetch').callsArgWith(1, null, { rows: [
-    { doc: place_contact },
-    { doc: place_parentContact },
-    { doc: place_grandparentContact }
-  ] });
-  lineage.fetchHydratedDoc(place._id).then(actual => {
-    test.equals(actual.name, place.name);
-    test.equals(actual.contact.phone, '+789');
-    test.equals(actual.parent.name, place_parent.name);
-    test.equals(actual.parent.contact.phone, '+123');
-    test.equals(actual.parent.parent.name, place_grandparent.name);
-    test.equals(actual.parent.parent.contact.phone, '+456');
-    test.equals(view.callCount, 1);
-    test.equals(view.args[0][0], 'medic-client');
-    test.equals(view.args[0][1], 'docs_by_id_lineage');
-    test.deepEqual(view.args[0][2].startkey, [ place._id ]);
-    test.deepEqual(view.args[0][2].endkey, [ place._id, {} ]);
-    test.equals(view.args[0][2].include_docs, true);
-    test.equals(fetch.callCount, 1);
-    test.deepEqual(fetch.args[0][0].keys, [ place_contact._id, place_parentContact._id, place_grandparentContact._id ]);
-    test.done();
-  }).catch(test.done);
-};
-
-// This is a classic use-case: report from CHW who is the contact for their own area
-exports['fetchHydratedDoc attaches re-used contacts, minify handles the circular references'] = test => {
-  const docId = 'docId';
-  const chwId = 'chwId';
-  const areaId = 'areaId';
-
-  const doc = {type: 'data_record', _id: docId, contact: {_id: chwId, parent: {_id: areaId}}};
-  const chw = {type: 'person', _id: chwId, parent: {_id: areaId}, hydrated: true};
-  const area = {type: 'clinic', _id: areaId, contact: {_id: chwId }};
-
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: doc},
-    { doc: chw},
-    { doc: area},
-  ]});
-  sinon.stub(db.medic, 'fetch').callsArgWith(1, Error('Calling fetch should not be required'));
-  lineage.fetchHydratedDoc(docId).then(actual => {
-    // The contact and the contact's parent's contact are the hydrated CHW
-    test.equal(actual.contact.parent.contact.hydrated, true);
-    test.equal(actual.contact.hydrated, true);
-
-    // And we can minifiy back to the original without error
-    lineage.minify(actual);
-    test.deepEqual(actual, doc);
-
-    test.done();
-  }).catch(test.done);
-};
-
-// Same as above but expanded to need to pull _some_ contacts
-exports['fetchHydratedDoc only fetches contacts that it has not got via lineage'] = test => {
-  const docId = 'docId';
-  const chwId = 'chwId';
-  const areaId = 'areaId';
-  const dhId = 'dhId';
-  const dhContactId = 'dhContactId';
-
-  const doc = {type: 'data_record', _id: docId, contact: {_id: chwId, parent: {_id: areaId, parent: {_id: dhId}}}};
-  const chw = {type: 'person', _id: chwId, parent: {_id: areaId, parent: {_id: dhId}}, hydrated: true};
-  const area = {type: 'clinic', _id: areaId, contact: {_id: chwId }, parent: {_id: dhId}};
-  const dh = {type: 'district-hospital', _id: dhId, contact: {_id: dhContactId}};
-  const dhContact = {type: 'person', _id: dhContactId, hydrated: true};
-
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: doc},
-    { doc: chw},
-    { doc: area},
-    { doc: dh}
-  ]});
-  const fetchStub = sinon.stub(db.medic, 'fetch');
-  fetchStub.callsArgWith(1, null, { rows: [{doc: dhContact}]});
-  lineage.fetchHydratedDoc(docId).then(actual => {
-    test.equal(actual.contact.hydrated, true);
-    test.equal(actual.contact.parent.contact.hydrated, true);
-    test.equal(actual.contact.parent.parent.contact.hydrated, true);
-
-    test.deepEqual(fetchStub.args[0][0], {keys: [dhContactId]});
-
-    test.done();
-  }).catch(test.done);
-};
-
-exports['minify handles null argument'] = test => {
-  lineage.minify(null);
-  // just make sure it doesn't blow up!
-  test.done();
-};
-
-exports['minify minifies the parent'] = test => {
-  const actual = {
-    _id: 'c',
-    name: 'cathy',
-    parent: {
-      _id: 'a',
-      name: 'arnold',
-      parent: {
-        _id: 'b',
-        name: 'barry'
-      }
-    }
-  };
-  const expected = {
-    _id: 'c',
-    name: 'cathy',
-    parent: {
-      _id: 'a',
-      parent: {
-        _id: 'b'
-      }
-    }
-  };
-  lineage.minify(actual);
-  test.deepEqual(actual, expected);
-  test.done();
-};
-
-exports['minify minifies the contact and lineage'] = test => {
-  const actual = {
-    _id: 'c',
-    name: 'cathy',
-    parent: {
-      _id: 'a',
-      name: 'arnold',
-      parent: {
-        _id: 'b',
-        name: 'barry'
-      }
-    },
-    contact: {
-      _id: 'd',
-      name: 'daniel',
-      parent: {
-        _id: 'e',
-        name: 'elisa'
-      }
-    }
-  };
-  const expected = {
-    _id: 'c',
-    name: 'cathy',
-    parent: {
-      _id: 'a',
-      parent: {
-        _id: 'b'
-      }
-    },
-    contact: {
-      _id: 'd',
-      parent: {
-        _id: 'e'
-      }
-    }
-  };
-  lineage.minify(actual);
-  test.deepEqual(actual, expected);
-  test.done();
-};
-
-
-exports['minify removes the patient'] = test => {
-  const actual = {
-    _id: 'c',
-    type: 'data_record',
-    patient_id: '123',
-    patient: {
-      _id: 'a',
-      name: 'Alice',
-      patient_id: '123',
-      parent: {
-        _id: 'b',
-        name: 'Bob'
-      }
-    }
-  };
-  const expected = {
-    _id: 'c',
-    type: 'data_record',
-    patient_id: '123'
-  };
-  lineage.minify(actual);
-  test.deepEqual(actual, expected);
-  test.done();
-};
-
-exports['fetchHydratedDoc+minify is noop on a report'] = test => {
-  sinon.stub(utils, 'getPatientContactUuid').callsArgWith(2, null, report_patient._id);
-  const viewStub = sinon.stub(db.medic, 'view');
-  viewStub.onCall(0).callsArgWith(3, null, { rows: [
-    { doc: JSON.parse(JSON.stringify(report)) },
-    { doc: report_contact },
-    { doc: report_parent },
-    { doc: report_grandparent }
-  ] });
-  viewStub.onCall(1).callsArgWith(3, null, { rows: [
-    { doc: report_patient },
-    { doc: report_contact },
-    { doc: report_parent },
-    { doc: report_grandparent }
-  ] });
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.callsFake((options, callback) => {
-    const contactDocs = options.keys.map(id => {
-      return [ report_contact, report_parentContact, report_grandparentContact, report_patient ]
-        .find(contact => contact._id === id);
-    });
-    callback(null, { rows: contactDocs.map(doc => { return {doc: doc}; }) });
-  });
-
-  lineage.fetchHydratedDoc(report._id).then(actual => {
-    lineage.minify(actual);
-    test.deepEqual(actual, report);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc+minify is noop on a place'] = test => {
-  sinon.stub(db.medic, 'view').callsArgWith(3, null, { rows: [
-    { doc: JSON.parse(JSON.stringify(place)) },
-    { doc: place_parent },
-    { doc: place_grandparent }
-  ] });
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.callsFake((options, callback) => {
-    const contactDocs = options.keys.map(id => {
-      return [ place_contact, place_parentContact, place_grandparentContact ]
-        .find(contact => contact._id === id);
-    });
-    callback(null, { rows: contactDocs.map(doc => { return {doc: doc}; }) });
-  });
-
-  lineage.fetchHydratedDoc(place._id).then(actual => {
-    lineage.minify(actual);
-    test.deepEqual(actual, place);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['hydrateDocs binds contacts and parents'] = test => {
-  const docs = [ report, place ];
-
-  const fetchedParents = [ report_parent, report_grandparent, report_contact, place_parent, place_grandparent ];
-  const fetchedContacts = [ report_parentContact, report_grandparentContact, place_contact, place_parentContact, place_grandparentContact ];
-  const rowify = docs => ({ rows: docs.map(doc => ({ id: doc._id, doc: doc })) });
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.onCall(0).callsArgWith(1, null, rowify(fetchedParents));
-  fetch.onCall(1).callsArgWith(1, null, rowify(fetchedContacts));
-
-  sinon.stub(utils, 'getPatientContactUuid').callsArgWith(2, null, report_patient._id);
-  const viewStub = sinon.stub(db.medic, 'view');
-  viewStub.onCall(0).callsArgWith(3, null, { rows: [
-    { doc: report_patient }
-  ] });
-
-  lineage.hydrateDocs(docs).then(([ hydratedReport, hydratedPlace ]) => {
-    test.equals(fetch.callCount, 2);
-    test.deepEqual(fetch.args[0][0].keys.sort(), fetchedParents.map(doc => doc._id).sort());
-    test.deepEqual(fetch.args[1][0].keys.sort(), fetchedContacts.map(doc => doc._id).sort());
-
-    test.equals(hydratedReport.contact.name, report_contact.name);
-    test.equals(hydratedReport.parent, null);
-    test.equals(hydratedReport.contact.parent.name, report_parent.name);
-    test.equals(hydratedReport.contact.parent.contact.name, report_parentContact.name);
-    test.equals(hydratedReport.contact.parent.parent.name, report_grandparent.name);
-    test.equals(hydratedReport.contact.parent.parent.contact.name, report_grandparentContact.name);
-    test.equals(hydratedReport.patient._id, report_patient._id);
-
-    test.equals(hydratedPlace.contact.name, place_contact.name);
-    test.equals(hydratedPlace.parent.name, place_parent.name);
-    test.equals(hydratedPlace.parent.contact.name, place_parentContact.name);
-    test.equals(hydratedPlace.parent.parent.name, place_grandparent.name);
-    test.equals(hydratedPlace.parent.parent.contact.name, place_grandparentContact.name);
-
-    test.done();
-  }).catch(test.done);
-};
-
-exports['hydrateDocs works on empty array'] = test => {
-  const docs = [];
-  lineage.hydrateDocs(docs).then((hydratedDocs) => {
-    test.equals(hydratedDocs.length, 0);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['hydrateDocs works on docs without contacts or parents'] = test => {
-  const docs = [
-    { _id: 'a' },
-    { _id: 'b' },
-  ];
-  lineage.hydrateDocs(docs).then((hydratedDocs) => {
-    test.deepEqual(hydratedDocs, docs);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['hydrateDocs+minify is noop'] = test => {
-  const docs = [ report, place ];
-
-  const fetchedParents = [ report_parent, report_grandparent, report_contact, place_parent, place_grandparent ];
-  const fetchedContacts = [ report_parentContact, report_grandparentContact, place_contact, place_parentContact, place_grandparentContact ];
-  const rowify = docs => ({ rows: docs.map(doc => ({ id: doc._id, doc: doc })) });
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.onCall(0).callsArgWith(1, null, rowify(fetchedParents));
-  fetch.onCall(1).callsArgWith(1, null, rowify(fetchedContacts));
-
-  sinon.stub(utils, 'getPatientContactUuid').callsArgWith(2, null, report_patient._id);
-  const viewStub = sinon.stub(db.medic, 'view');
-  viewStub.onCall(0).callsArgWith(3, null, { rows: [
-    { doc: report_patient }
-  ] });
-
-  lineage.hydrateDocs(docs).then(([ hydratedReport, hydratedPlace ]) => {
-    lineage.minify(hydratedReport);
-    lineage.minify(hydratedPlace);
-    test.deepEqual(hydratedReport, report);
-    test.deepEqual(hydratedPlace, place);
-    test.done();
-  }).catch(test.done);
-};
-
-exports['hydrateDocs ignores db-fetch errors'] = test => {
-  const docs = [ report, place ];
-
-  const fetchedParents = [ report_parent, report_grandparent, report_contact, place_parent, place_grandparent ];
-  const fetchedContacts = [ report_parentContact, report_grandparentContact, place_contact, place_parentContact, place_grandparentContact ];
-  const rowify = docs => ({ rows: docs.map(doc => ({ id: doc._id, doc: doc })) });
-
-  const fetchedParentsRows = rowify(fetchedParents);
-  fetchedParentsRows.rows[3].doc = null; // place_parent
-  const fetchedContactsRows = rowify(fetchedContacts);
-  fetchedContactsRows.rows[0].doc = null; // report_parentContact
-
-  const fetch = sinon.stub(db.medic, 'fetch');
-  fetch.onCall(0).callsArgWith(1, null, fetchedParentsRows);
-  fetch.onCall(1).callsArgWith(1, null, fetchedContactsRows);
-
-  sinon.stub(utils, 'getPatientContactUuid').callsArgWith(2, null, report_patient._id);
-  const viewStub = sinon.stub(db.medic, 'view');
-  viewStub.onCall(0).callsArgWith(3, null, { rows: [
-    { doc: report_patient }
-  ] });
-
-  lineage.hydrateDocs(docs).then(([ hydratedReport, hydratedPlace ]) => {
-    test.equals(hydratedReport.contact.name, report_contact.name);
-    test.equals(hydratedReport.parent, null);
-    test.equals(hydratedReport.contact.parent.name, report_parent.name);
-    test.equals(hydratedReport.contact.parent.contact._id, report_parentContact._id);
-    test.equals(hydratedReport.contact.parent.contact.name, null); // db-fetch error : not hydrated
-    test.equals(hydratedReport.contact.parent.parent.name, report_grandparent.name);
-    test.equals(hydratedReport.contact.parent.parent.contact.name, report_grandparentContact.name);
-
-    test.equals(hydratedPlace.contact.name, place_contact.name);
-    test.equals(hydratedPlace.parent._id, place_parent._id);
-    test.equals(hydratedPlace.parent.name, null); // db-fetch error : not hydrated
-    test.equals(hydratedPlace.parent.contact, null); // db-fetch error : not hydrated
-    test.equals(hydratedPlace.parent.parent.name, place_grandparent.name);
-    test.equals(hydratedPlace.parent.parent.contact.name, place_grandparentContact.name);
-
-    test.done();
-  }).catch(test.done);
-};
-
-exports['fetchHydratedDoc works for SMS reports'] = test => {
-  const view = sinon.stub(db.medic, 'view');
-  let doc = {
-    _id: 'some_id',
-    type: 'data_record',
-    from: '123',
-    form: 'D',
-    fields: {},
-    sms_message: {
-      message_id: '76992',
-      message: 'somemessage',
-      from: '123',
-      type: 'sms_message',
-      form: 'D'
-    }
-  };
-
-  view.onCall(0).callsArgWith(3, null, { rows: [{ doc: doc }] });
-  lineage
-    .fetchHydratedDoc(doc._id)
-    .then(() => {
-      test.equals(view.callCount, 1);
-      test.done();
-    })
-    .catch(err => {
-      test.equals(err instanceof TypeError, true);
-      test.equals(err.message, 'Cannot read property \'parent\' of undefined');
-      test.equals(view.callCount, 1);
-      test.ok(false, 'fails with error: '+ err.message);
-      test.done();
-    });
-
-};
