+++ shared-libs/lineage/test/.jshintrc
@@ -1,4 +1,5 @@
 {
   "extends": "../../.jshintrc",
-  "esversion": 6
+  "esversion": 6,
+  "expr": true
 }
